fix: Jim found a bug with routing slip.
Basically the Rest component predated when Vramel fully ported most of
camel and therefore it didn't do everything right by camel. It wasn't
using a unique endpoint url which messed up the producer cache. This
caused a problem with the routingslip because the cachekey didn't take
into account changes to host, etc.